+++ modules/runners/outputs.tf
@@ -10,6 +10,10 @@ output "lambda_scale_up" {
   value = aws_lambda_function.scale_up
 }
 
+output "lambda_scale_up_log_group" {
+  value = aws_cloudwatch_log_group.scale_up
+}
+
 output "role_scale_up" {
   value = aws_iam_role.scale_up
 }
@@ -18,12 +22,29 @@ output "lambda_scale_down" {
   value = aws_lambda_function.scale_down
 }
 
+output "lambda_scale_down_log_group" {
+  value = aws_cloudwatch_log_group.scale_down
+}
+
 output "role_scale_down" {
   value = aws_iam_role.scale_down
 }
 
+output "lambda_pool" {
+  value = try(module.pool[0].lambda, null)
+}
+
+output "lambda_pool_log_group" {
+  value = try(module.pool[0].lambda_log_group, null)
+}
+
 output "role_pool" {
-  value = length(var.pool_config) == 0 ? null : module.pool[0].role_pool
+  value = try(module.pool[0].role_pool, null)
+}
+
+output "runners_log_groups" {
+  description = "List of log groups from different log files of runner machine."
+  value       = try(aws_cloudwatch_log_group.gh_runners, [])
 }
 
 output "logfiles" {