+++ modules/runners/pool.tf
@@ -15,9 +15,9 @@ module "pool" {
     instance_target_capacity_type = var.instance_target_capacity_type
     instance_types                = var.instance_types
     kms_key_arn                   = local.kms_key_arn
+    ami_kms_key_arn               = local.ami_kms_key_arn
     lambda = {
       log_level                      = var.log_level
-      log_type                       = var.log_type
       logging_retention_in_days      = var.logging_retention_in_days
       logging_kms_key_id             = var.logging_kms_key_id
       reserved_concurrent_executions = var.pool_lambda_reserved_concurrent_executions
@@ -41,12 +41,15 @@ module "pool" {
       extra_labels              = var.runner_extra_labels
       launch_template           = aws_launch_template.runner
       group_name                = var.runner_group_name
+      name_prefix               = var.runner_name_prefix
       pool_owner                = var.pool_runner_owner
       role                      = aws_iam_role.runner
     }
-    subnet_ids     = var.subnet_ids
-    ssm_token_path = "${var.ssm_paths.root}/${var.ssm_paths.tokens}"
-    tags           = local.tags
+    subnet_ids                           = var.subnet_ids
+    ssm_token_path                       = "${var.ssm_paths.root}/${var.ssm_paths.tokens}"
+    ami_id_ssm_parameter_name            = var.ami_id_ssm_parameter_name
+    ami_id_ssm_parameter_read_policy_arn = var.ami_id_ssm_parameter_name != null ? aws_iam_policy.ami_id_ssm_parameter_read[0].arn : null
+    tags                                 = local.tags
   }
 
   aws_partition = var.aws_partition